Language.Drasil.Code.Expr.Development

Description

Re-exporting modules

Documentation

data CodeExpr where

Expression language where all terms also denote a term in GOOL (i.e. translation is total and meaning preserving).

Constructors

Lit :: Literal -> CodeExpr
Brings literals into the expression language.

AssocA :: AssocArithOper -> [CodeExpr] -> CodeExpr
Takes an associative arithmetic operator with a list of expressions.

AssocB :: AssocBoolOper -> [CodeExpr] -> CodeExpr
Takes an associative boolean operator with a list of expressions.

C :: UID -> CodeExpr
C stands for Chunk, for referring to a chunk in an expression. Implicitly assumes that the chunk has a symbol.

FCall :: UID -> [CodeExpr] -> [(UID, CodeExpr)] -> CodeExpr
A function call accepts a list of parameters and a list of named parameters. For example

F(x) is (FCall F [x] []).

F(x,y) would be (FCall F [x,y]).

F(x,n=y) would be (FCall F [x] [(n,y)]).

New :: UID -> [CodeExpr] -> [(UID, CodeExpr)] -> CodeExpr
Actor creation given UID, parameters, and named parameters.

Message :: UID -> UID -> [CodeExpr] -> [(UID, CodeExpr)] -> CodeExpr
Message an actor:

1st UID is the actor,

2nd UID is the method.

Field :: UID -> UID -> CodeExpr
Access a field of an actor:

1st UID is the actor,

2nd UID is the field.

Case :: Completeness -> [(CodeExpr, CodeExpr)] -> CodeExpr
For multi-case expressions, each pair represents one case.

Matrix :: [[CodeExpr]] -> CodeExpr
Represents a matrix of expressions.

UnaryOp :: UFunc -> CodeExpr -> CodeExpr
Unary operation for most functions (eg. sin, cos, log, etc.).

UnaryOpB :: UFuncB -> CodeExpr -> CodeExpr
Unary operation for Bool -> Bool operations.

UnaryOpVV :: UFuncVV -> CodeExpr -> CodeExpr
Unary operation for Vector -> Vector operations.

UnaryOpVN :: UFuncVN -> CodeExpr -> CodeExpr
Unary operation for Vector -> Number operations.

ArithBinaryOp :: ArithBinOp -> CodeExpr -> CodeExpr -> CodeExpr
Binary operator for arithmetic between expressions (fractional, power, and subtraction).

BoolBinaryOp :: BoolBinOp -> CodeExpr -> CodeExpr -> CodeExpr
Binary operator for boolean operators (implies, iff).

EqBinaryOp :: EqBinOp -> CodeExpr -> CodeExpr -> CodeExpr
Binary operator for equality between expressions.

LABinaryOp :: LABinOp -> CodeExpr -> CodeExpr -> CodeExpr
Binary operator for indexing two expressions.

OrdBinaryOp :: OrdBinOp -> CodeExpr -> CodeExpr -> CodeExpr
Binary operator for ordering expressions (less than, greater than, etc.).

VVVBinaryOp :: VVVBinOp -> CodeExpr -> CodeExpr -> CodeExpr
Binary operator for Vector x Vector -> Vector operations (cross product).

VVNBinaryOp :: VVNBinOp -> CodeExpr -> CodeExpr -> CodeExpr
Binary operator for Vector x Vector -> Number operations (dot product).

Operator :: AssocArithOper -> DiscreteDomainDesc CodeExpr CodeExpr -> CodeExpr -> CodeExpr
Operators are generalized arithmetic operators over a DomainDesc of an Expr. Could be called BigOp. ex: Summation is represented via Add over a discrete domain.

RealI :: UID -> RealInterval CodeExpr CodeExpr -> CodeExpr
The expression is an element of a space. IsIn :: Expr -> Space -> Expr | A different kind of IsIn. A UID is an element of an interval.

data ArithBinOp

Arithmetic operators (fractional, power, and subtraction).

Constructors

Frac

Pow

Subt

data EqBinOp

Equality operators (equal or not equal).

Constructors

Eq

NEq

data BoolBinOp

Conditional and Biconditional operators (Expressions can imply one another, or exist if and only if another expression exists).

Constructors

Impl

Iff

data LABinOp

Index operator.

Constructors

Index

data OrdBinOp

Ordered binary operators (less than, greater than, less than or equal to, greater than or equal to).

Constructors

Lt

Gt

LEq

GEq

data VVVBinOp

Vector x Vector -> Vector binary operations (cross product).

Constructors

Cross

data VVNBinOp

Vector x Vector -> Number binary operations (dot product).

Constructors

Dot

data AssocArithOper

Associative operators (adding and multiplication). Also specifies whether it is for integers or for real numbers.

Constructors

AddI

AddRe

MulI

MulRe

data AssocBoolOper

Associative boolean operators (and, or).

Constructors

And

Or

data UFunc

Unary functions (abs, log, ln, sin, etc.).

Constructors

Abs

Log

Ln

Sin

Cos

Tan

Sec

Csc

Cot

Arcsin

Arccos

Arctan

Exp

Sqrt

Neg

data UFuncB

Bool -> Bool operators.

Constructors

Not

data UFuncVV

Vector -> Vector operators.

Constructors

NegV

data UFuncVN

Vector -> Number operators.

Constructors

Norm

Dim

eDep :: CodeExpr -> [UID]

Get dependencies from an equation.

eDep' :: CodeExpr -> [UID]

Get dependencies from an equation, without functions.

eNamesRI :: RealInterval CodeExpr CodeExpr -> [UID]

Generic traversal of everything that could come from an interval to names (similar to eNames).

eNamesRI' :: RealInterval CodeExpr CodeExpr -> [UID]

Generic traversal of everything that could come from an interval to names without functions (similar to eNames').

eprec :: CodeExpr -> Int

eprec - Expression precedence.

precA :: AssocArithOper -> Int

precA - precedence for arithmetic-related Binary-Associative (Commutative) operators.

precB :: AssocBoolOper -> Int

precB - precedence for boolean-related Binary-Associative (Commutative) operators.

expr :: Expr -> CodeExpr

Render an algebraic expression into our code expression language.

realInterval :: RealInterval Expr Expr -> RealInterval CodeExpr CodeExpr

Convert RealInterval Expr Exprs into RealInterval CodeExpr CodeExprs.

constraint :: ConstraintE -> Constraint CodeExpr

Convert constrained expressions (ConstraintE) into Constraint''CodeExprs.