Language.Drasil.CodeExpr

Documentation

data CodeExpr

A one-to-one clone of Expr, with extra OO/code-related functionality.

sy :: (HasUID u, HasSymbol u) => u -> CodeExpr

Smart constructor for chunk symbols.

str :: String -> CodeExpr

Smart constructor for strings.

int :: Integer -> CodeExpr

Smart constructor for integers.

dbl :: Double -> CodeExpr

Smart constructor for doubles.

exactDbl :: Integer -> CodeExpr

Smart constructor for exact doubles.

matrix :: [[CodeExpr]] -> CodeExpr

Smart constructor for matrices.

frac :: Integer -> Integer -> CodeExpr

Smart constructor for fractions.

recip_ :: CodeExpr -> CodeExpr

Smart constructor for rational expressions (only in 1/x form).

abs_ :: CodeExpr -> CodeExpr

Smart constructor for taking the absolute value of an expression.

neg :: CodeExpr -> CodeExpr

Smart constructor for negating an expression.

log :: CodeExpr -> CodeExpr

Smart constructor to take the log of an expression.

ln :: CodeExpr -> CodeExpr

Smart constructor to take the ln of an expression.

sqrt :: CodeExpr -> CodeExpr

Smart constructor to take the square root of an expression.

sin :: CodeExpr -> CodeExpr

Smart constructor to apply sin to an expression.

cos :: CodeExpr -> CodeExpr

Smart constructor to apply cos to an expression.

tan :: CodeExpr -> CodeExpr

Smart constructor to apply tan to an expression.

sec :: CodeExpr -> CodeExpr

Smart constructor to apply sec to an expression.

csc :: CodeExpr -> CodeExpr

Smart constructor to apply csc to an expression.

cot :: CodeExpr -> CodeExpr

Smart constructor to apply cot to an expression.

arcsin :: CodeExpr -> CodeExpr

Smart constructor to apply arcsin to an expression.

arccos :: CodeExpr -> CodeExpr

Smart constructor to apply arccos to an expression.

arctan :: CodeExpr -> CodeExpr

Smart constructor to apply arctan to an expression.

exp :: CodeExpr -> CodeExpr

Smart constructor for the exponential (base e) function.

dim :: CodeExpr -> CodeExpr

Smart constructor for calculating the dimension of a vector.

norm :: CodeExpr -> CodeExpr

Smart constructor for calculating the normal form of a vector.

not_ :: CodeExpr -> CodeExpr

Smart constructor for applying logical negation to an expression.

new :: (Callable f, HasUID f, CodeIdea f) => f -> [CodeExpr] -> CodeExpr

Constructs a CodeExpr for actor creation (constructor call)

newWithNamedArgs :: (Callable f, HasUID f, CodeIdea f, HasUID a, IsArgumentName a) => f -> [CodeExpr] -> [(a, CodeExpr)] -> CodeExpr

Constructs a CodeExpr for actor creation (constructor call) that uses named arguments

message :: (Callable f, HasUID f, CodeIdea f, HasUID c, HasSpace c, CodeIdea c) => c -> f -> [CodeExpr] -> CodeExpr

Constructs a CodeExpr for actor messaging (method call)

msgWithNamedArgs :: (Callable f, HasUID f, CodeIdea f, HasUID c, HasSpace c, CodeIdea c, HasUID a, IsArgumentName a) => c -> f -> [CodeExpr] -> [(a, CodeExpr)] -> CodeExpr

Constructs a CodeExpr for actor messaging (method call) that uses named arguments

field :: CodeVarChunk -> CodeVarChunk -> CodeExpr

Constructs a CodeExpr representing the field of an actor

apply :: (HasUID f, HasSymbol f) => f -> [CodeExpr] -> CodeExpr

Applies a given function with a list of parameters.

apply1 :: (HasUID f, HasSymbol f, HasUID a, HasSymbol a) => f -> a -> CodeExpr

Similar to apply, but converts second argument into Symbols.

apply2 :: (HasUID f, HasSymbol f, HasUID a, HasSymbol a, HasUID b, HasSymbol b) => f -> a -> b -> CodeExpr

Similar to apply, but the applied function takes two parameters (which are both Symbols).

applyWithNamedArgs :: (HasUID f, HasSymbol f, HasUID a, IsArgumentName a) => f -> [CodeExpr] -> [(a, CodeExpr)] -> CodeExpr

Similar to apply, but takes a relation to apply to FCall.

($&&) :: CodeExpr -> CodeExpr -> CodeExpr

Smart constructor for the boolean and operator.

($-) :: CodeExpr -> CodeExpr -> CodeExpr

Smart constructor for subtracting two expressions.

($/) :: CodeExpr -> CodeExpr -> CodeExpr

Smart constructor for dividing two expressions.

($^) :: CodeExpr -> CodeExpr -> CodeExpr

Smart constructor for rasing the first expression to the power of the second.

($=>) :: CodeExpr -> CodeExpr -> CodeExpr

Smart constructor to show that one expression implies the other (conditional operator).

($<=>) :: CodeExpr -> CodeExpr -> CodeExpr

Smart constructor to show that an expression exists if and only if another expression exists (biconditional operator).

($=) :: CodeExpr -> CodeExpr -> CodeExpr

Smart constructor for equating two expressions.

($!=) :: CodeExpr -> CodeExpr -> CodeExpr

Smart constructor for showing that two expressions are not equal.

($<) :: CodeExpr -> CodeExpr -> CodeExpr

Less than.

($>) :: CodeExpr -> CodeExpr -> CodeExpr

Greater than.

($<=) :: CodeExpr -> CodeExpr -> CodeExpr

Less than or equal to.

($>=) :: CodeExpr -> CodeExpr -> CodeExpr

Greater than or equal to.

addI :: CodeExpr -> CodeExpr -> CodeExpr

Add two expressions (Integers).

addRe :: CodeExpr -> CodeExpr -> CodeExpr

Add two expressions (Real numbers).

mulI :: CodeExpr -> CodeExpr -> CodeExpr

Multiply two expressions (Integers).

mulRe :: CodeExpr -> CodeExpr -> CodeExpr

Multiply two expressions (Real numbers).

idx :: CodeExpr -> CodeExpr -> CodeExpr

Smart constructor for indexing.

expr :: Expr -> CodeExpr

Render an algebraic expression into our code expression language